scanner/app/scanner.py
import nmap
import os
import requests
import time

class BasicScanner:
    def __init__(self):
        self.nmap_scanner = nmap.PortScanner()
        self.api_url = os.getenv("API_URL", "http://api:8000")

    def scan_network(self, target):
        """Basic network scan"""
        return self.nmap_scanner.scan(target, arguments='-F --script=vuln')

def main():
    scanner = BasicScanner()
    while True:
        # Add scanning logic here
        time.sleep(60)

if __name__ == "__main__":
    main()
