scanner/requirements.txt
requests==2.31.0
python-nmap==0.7.1
python-dotenv==1.0.0
